requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
black==23.7.0
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.4
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
docker-pycreds==0.4.0
einops==0.6.1
filelock==3.11.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
fvcore==0.1.5.post20221221
gitdb==4.0.10
gitpython==3.1.31
gmpy2==2.1.2
google-auth-oauthlib==1.0.0
google-auth==2.17.3
grad-cam==1.4.8
grpcio==1.54.0
h5py==3.8.0
huggingface-hub==0.13.4
hydra-core==1.3.2
idna==3.4
imagecorruptions==1.1.2
imageio==2.31.1
importlib-metadata==6.5.0
importlib-resources==5.12.0
iopath==0.1.9
jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy-loader==0.3
lightning-utilities==0.8.0
lit==16.0.6
markdown-it-py==3.0.0
markdown==3.4.3
markupsafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
munkres==1.1.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.68
packaging==23.1
pandas==2.0.0
pathspec==0.11.2
pathtools==0.1.2
pillow==9.4.0
pip==22.3.1
platformdirs==3.2.0
ply==3.11
pooch==1.7.0
portalocker==2.7.0
protobuf==4.22.3
psutil==5.9.5
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycocotools==2.0.6
pycparser==2.21
#pydensecrf==1.0rc2
pygments==2.16.1
pyopenssl==23.1.1
pyparsing==3.0.9
pyqt5-sip==12.11.0
pyqt5==5.15.7
pysocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==2.0.1.post0
pytz==2023.3
pywavelets==1.4.1
pyyaml==6.0
requests-oauthlib==1.3.1
requests==2.28.2
rich==13.6.0
rsa==4.9
safetensors==0.4.0
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.19.1
setproctitle==1.3.2
setuptools==67.6.1
shapely==2.0.1
sip==6.7.8
six==1.16.0
smmap==5.0.0
sympy==1.11.1
tabulate==0.9.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboard==2.12.2
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.8.12
timm==0.9.0
toml==0.10.2
tomli==2.0.1
torch-geometric==2.3.1
#torch==2.1.0
#torchaudio==2.1.0
torchmetrics==0.11.4
#torchvision==0.16.0
tornado==6.2
tqdm==4.65.0
triton==2.1.0
ttach==0.0.3
typing-extensions==4.5.0
tzdata==2023.3
unicodedata2==15.0.0
urllib3==1.26.15
wandb==0.14.2
werkzeug==2.2.3
wget==3.2
wheel==0.40.0
yacs==0.1.8
yarl==1.8.2
zipp==3.15.0